Ported to snowpack for hot-module replacement and npm package management
