+++ public/assets/hand.png
