+++ public/index.html
@@ -3,18 +3,13 @@
 <head>
 	<meta charset="utf-8">
 	<title>Bongo Grapher</title>
-	<link rel="stylesheet" href="./stylesheet.css"/>
+	<link rel="stylesheet" href="/stylesheet.css"/>
 </head>
 
 <body>
 	<center id = "body">
 		<h1 id="title">Bongo Cat Grapher</h1>
-		<script src =  "./evaluatex.min.js"></script>
-		<script>
-			evaluatex();
-		</script>
-		<!--script src="https://www.desmos.com/api/v1.5/calculator.js?apiKey=dcb31709b452b1cf9dc26972add0fda6"></script-->
-		<script src = "./grapher.js"></script>
+    <script type="module" src="/_dist_/index.js"></script>
 		<canvas id = "canvas" width="1000" height = "720"></canvas>
 		
 		<table id = "compact">
