+++ snowpack.config.json
@@ -1,0 +1,7 @@
+{
+  "scripts": {
+    "mount:public": "mount public --to /",
+    "mount:src": "mount src --to /_dist_"
+  },
+  "plugins": []
+}
