+++ snowpack.lock.json
@@ -1,0 +1,4 @@
+{
+  "imports": {
+  }
+}
